requirements.txt
python-telegram-bot==20.*
pexpect
python-dotenv